-Adding Data SQL File -Fix Setup and Audit it, running. -Test of Relative Paths, runnning.

+++ login.php
@@ -25,13 +25,22 @@ if(isset($_POST['username']) && isset($_POST['password'])) {
             <p><input type="submit" value="Anmelden"/></p>
 </form>
 <?php 
-if (isset($_GET['message_type']) && isset($_GET['message_reason'])) {
-$type = $_GET['message_type'];
-$reason = $_GET['message_reason'];
-echo "<div class='login_msgbox'>";
-displayMessage($type,$reason);
-echo "</div>";
+if (file_exists('./setup/index.php')) {
+    $msgtype = "notice";
+    $msgreason = "setupfolder";
+    echo "<div class='login_msgbox'>";
+    displayMessage($msgtype,$msgreason);
+    echo "</div>";
+} else {
+    if (isset($_GET['message_type']) && isset($_GET['message_reason'])) {
+        $type = $_GET['message_type'];
+        $reason = $_GET['message_reason'];
+        echo "<div class='login_msgbox'>";
+        displayMessage($type,$reason);
+        echo "</div>";
+    }
 }
+
 ?></div>
 <div id="footer">
     <p>Copyright &copy; <?php echo date("Y"); ?> by <a href="www.pierewoehl.de">Piere W&ouml;hl</a> und <a href="mailto:manuel.lampe@hotmail.de">Manuel Lampe</a> Projektseite <a href="http://www.screenmediacast.net/projekte/?p=56">www.screenmediacast.net</a></p>
